requirements.txt
fastapi
uvicorn[standard]
streamlit
pandas
numpy
pymongo
tushare
akshare
requests
python-dotenv
apscheduler
aiohttp
pydantic
